scripts/requirements.txt
requests
jinja2
matplotlib
python-dotenv